fix crash for subcribe sa

+++ wifi/frameworks/native/c_adapter/src/wifi_c_event.cpp
@@ -329,10 +329,10 @@ void EventManager::Init()
     if (mSaStatusListener == nullptr) {
         WIFI_LOGI("EventManager Listener Init!");
         mSaStatusListener = new OHOS::Wifi::WifiAbilityStatusChange();
-        mSaStatusListener->Init(WIFI_DEVICE_ABILITY_ID);
-        mSaStatusListener->Init(WIFI_SCAN_ABILITY_ID);
-        mSaStatusListener->Init(WIFI_HOTSPOT_ABILITY_ID);
-        mSaStatusListener->Init(WIFI_P2P_ABILITY_ID);
+        mSaStatusListener->Init((int32_t)WIFI_DEVICE_ABILITY_ID);
+        mSaStatusListener->Init((int32_t)WIFI_SCAN_ABILITY_ID);
+        mSaStatusListener->Init((int32_t)WIFI_HOTSPOT_ABILITY_ID);
+        mSaStatusListener->Init((int32_t)WIFI_P2P_ABILITY_ID);
     }
     return;
 }
